pyproject.toml
[project]
name = "pima-docs"
version = "0.1.0"
description = "Gestão com IA de documentos incorporados a procedimentos da Promotoria"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.80.11",
    "pypdf2>=3.0.1",
    "pyside6>=6.10.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]

[tool.uv]
environments = ["sys_platform == 'win32' and platform_machine == 'AMD64'"]
include = ["main.py", "ui/**", "resources/**"]
